Replace failtag's magic numbers with the real enum
Correcting something I got wrong when I shipped failtag.c. It mapped Stripe
result codes by literal value -- case 1 for STRIPE_ERR_NETWORK, case 2 for
STRIPE_ERR_TLS -- with the names only in comments, to keep the file free of
stripe_api.h and therefore host-testable.

I said at the time that the correspondence was "pinned by test rather than
trusted". It was not. test_failtag hardcoded the same literals, so the test
asserted that the function did what I had written, not that what I had
written matched the enum. Reordering stripe_result_t would have left test
and implementation agreeing with each other and both wrong -- a silent
failure that reads as consistent in review.

Fixed at the root: stripe_result.h now holds the enum with no ESP-IDF
dependency, stripe_api.h includes it, and both failtag.c and its test use
the symbols. One definition, resolved by the compiler on both sides.

Switching on the enum with no default is the part that actually protects
this. Adding a value to stripe_result.h now fails the build with
-Werror=switch rather than silently falling through to "no tag" -- verified
by adding a STRIPE_ERR_TIMEOUT and watching it fail to compile. That is
better than a test, because it cannot be forgotten.

Also records two memory entries: shared assumptions between test and code
prove nothing, and the header-splitting pattern that keeps this codebase
host-testable.

21 suites, 1325 checks, 0 failures. Flashed and verified.

+++ firmware/main/main.c
@@ -853,7 +853,7 @@ static void refresh_task(void *arg)
             } else {
                 /* Remember why, so the stale footer can say whether the fault
                  * looks local or upstream. */
-                s_last_failure = failtag_from_result((int)r);
+                s_last_failure = failtag_from_result(r);
                 ESP_LOGW(TAG, "full refresh failed: %s", stripe_result_str(r));
             }
         }
